backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
drf-spectacular>=0.26.0
gunicorn>=20.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
pdf2image>=1.16.0
PyMuPDF>=1.21.1